+++ src/main/java/janggi/domain/piece/Camp.java
@@ -1,4 +1,4 @@
-package janggi.domain;
+package janggi.domain.piece;
 
 public enum Camp {
     HAN, CHO
